+++ CHANGELOG.md
@@ -28,6 +28,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Message Component
 - Header Component
 - Menu Component
+- Cards Component
+- Labels Component
 
 ### Changed